Removes unused code and makes components modular
