+++ react/src/components/presentation/Box.css
@@ -2,6 +2,7 @@
     position: absolute;
     outline: 5px solid;
     box-shadow: 0 0 10px black;
+    /* padding: 10px; */
 }
 
 .face-box:hover {
