+++ react/src/components/presentation/MaskResults.css
@@ -1,10 +1,11 @@
-.mask-results {
+.mask-detector-results {
     position: relative;
     max-height: 500px;
     max-width: 500px;
+    margin: 10px;
 }
 
-.mask-results img {
-    width: 100%;
-    height: 100%;
+.mask-detector-results__img {
+    max-width: 100%;
+    max-height: 100%;
 }
